Created interface between ObjectStorage and Replication, and moved files around
