+++ core/src/edu/rice/comp413fall2020grey/Common/Message.java
@@ -1,0 +1,5 @@
+package edu.rice.comp413fall2020grey.Common;
+
+public interface Message {
+
+}
